Fixed string int comparison in describe_sh_type(x)
When a numeric value of a section type is defined in the ENUM_SH_TYPE dict, the variable x is of a type string. This broke the tests when the section type was SHT_GNU_ATTRIBUTES, because it doesn't exist in the _DESCR_SH_TYPE dict, therefore it gets compared with an integer.