made elf/segments.py not choke on unterminated note names
elf/segments.py tries to parse ELF note names as CStrings. This will fail if the note name does not end in a NUL character. It turns out some binaries have note names that are not NUL-terminated. To prevent elftools from choking on such files, we detect when the name is not NUL-terminated and return all its bytes instead of trying to parse it as a CString.