Extend relocation patching in Relocationhandler
This adds two optional arguments to apply_section_relocations:

- symbol_mapping: The symbol mapping maps symbol ordinals to
addresses. If no symbol mapping is provided, symbols are taken to be at
absolute address st_value from their entry in the symbol table (as
was the behavior before).

- section_base can be passed, which is the starting
address of the section to be relocated in memory. This is used to
correctly emit pc-relative relocations. If not passed this will
default to 0.