Support for upstream construct 2.8.10
This commit provides the change from the 'old' construct library
to the new API introduced with version 2.8. Most of the changes are
names of data types (ULInt8 -> Int8ul and the like) and changing
over to the nameless initialization and construction using the '/'
operator.

The {U,S}LEB128 constructors have not yet been converted into the
new format so they still need the old syntax.

Two changes were required at places where If() was used as it a) lost
its 'elsevalue' parameter (requiring to move over to an IfThenElse
construct) and b) does not propagate Embedded() information anymore
(requiring to move Embedded outside the IfThenElse()).