Remove copy of construct 2.8.10, use dependency in setup.py
With the migration to the construct 2.8.X API done, we can also
remove the import of the current version from our tree and use
global imports for the required classes.

This change, however, currently makes it necessary to run pip install
with --process-dependency-links to install pyelftools correctly.

This is due to the fact that we can't just use the latest release
(2.8.10) of construct as there was a python2 incompatibility bug which
was only fixed later (see upstream commit 98d6c5fb50bc). Additionally,
the latest master revision (commit 4836722526cb) also changed the
predicate parameter to RepeatUntil in a backward-incompatible way
(which we use in dwarf/structs.py - this change should probably be
taken into our own RepeatUntilExcluding class as well).

Maybe there is a better way to resolve the dependency for now which I
don't know of, or we could wait for a new release of construct
(including the changes to RepeatUntil) which could then easily be
fetched from PyPI.