dwarf: DWARF v4 Type Unit support
Support for DWARF v4 Type Units and sig8 references.  DIE attributes
with form DW_FORM_ref_sig8 reference DIEs stored in the Type Unit with
the matching 8 byte signature in its header.  The specific DW_TAG_type
DIE referenced is also specified in the Type Unit header.  In DWARFv4
Type Units are collected in the in the .debug_types section.

CompleUnit objects for all Type Units in the .debug_types section are
created on the first request to iterate or search Type Units and are
cached in an OrderedDict keyed by the 64 bit signature as an integer.
The contained DIEs are parsed on demand like other CompileUnits.

The CompleUnit class (constructed with the header member containing the
TU header struct which has additional fields) is reused for Type Units.
Looking forward to DWARF v5 there are multiple unit types; each unit
type has a similar but customized header and all will be stored in the
.debug_info section.

readelf.py dumps the debug_types section after the debug_info section
if --dump-debug=info is selected, matching the readelf binary.  Also
the value description for DW_FORM_ref_sig8 attributes now matches
the readelf binary.