test file for dwarf macinfo vendor and partial units
This started as a unit test to make a vendor_ext for the
macinfo section.  The readelf tests were not passing with
just mentioning .debug_info and having the .debug_macinfo
section.  First readelf complained of no text section, then
an empty section choked up readelf.py.  Then readelf.py
wanted the line program, and the assember was inserting
a stub one that was choking readelf.