dwarf/structs : Single instance cache for DWARFStructs class
Cache the DWARFStructs class for each set of instantiation
parameters and returned the cache entry on future lookups.

Each compile unit parse requires two instances of DWARFStructs,
one default based on what is known of the elf file and a second
with the actual dwarf version of the compile unit.  Most of the
time these will result in duplicate classes.

Doing this conversion in just dwarfinfo cut the time to run all
unit tests from 13.263 to 12.933 seconds, a savings of 2.5%.