construct: typing: singletons
`NoDefault`, `Pass` and `Terminator` are all singletons, which are used
in various dictionaries. For type hinting the *type* is also required in
addition to the singleton *instance*.

Prefix the class with an underscore.

Hint instances with `Final[Any]` to prevent re-assignment and to exempt
them from type checking: Otherwise all code de-referencing something
from `dict[str, … | _Pass]` has to explicitly check for `_Pass` all
time, which currently it does not. Actually that is a bug, which needs
fixing at a later day.

Signed-off-by: Philipp Hahn <phahn-oss@avm.de>