Last typing issues (#610)
* Mark files as (not) executable

Mark data- and text files as not executable.
Mark scripts as executable.
Mark main tests runners as executable.

- https://docs.astral.sh/ruff/rules/#flake8-executable-exe

Signed-off-by: Philipp Hahn <phahn-oss@avm.de>

* Convert DW_OP_opcode2name into a dict-comprehension

Signed-off-by: Philipp Hahn <phahn-oss@avm.de>

* callframe: Simplify default value

Signed-off-by: Philipp Hahn <phahn-oss@avm.de>

* elf.descriptions: Simplify range check

Signed-off-by: Philipp Hahn <phahn-oss@avm.de>

* DWARFInfo: Remove inner self

`resolve_strings()` is an internal helper function and does not need
`self`.

Signed-off-by: Philipp Hahn <phahn-oss@avm.de>

* Fix DWARFInfo._parse_line_program_at_offset bytes

`replace_value()` works on `bytes`, as such encode the string using the
default `utf-8` encoding.

Signed-off-by: Philipp Hahn <phahn-oss@avm.de>

* Turn TypeDesc.dimensions into a tuple

`(… for … in …)` is a generator expression, which only can be walked
once compared to a `tuple`.

Signed-off-by: Philipp Hahn <phahn-oss@avm.de>

* Fix return type of _parse_cie_augmentation()

`elftools.dwarf.callframe.CallFrameInfo._parse_cie_augmentation()`
returns a `tuple[bytes, …]`.

Signed-off-by: Philipp Hahn <phahn-oss@avm.de>

---------

Signed-off-by: Philipp Hahn <phahn-oss@avm.de>