Handle hash sections with length 0
When parsing an ELF file with an empty hash section (length 0), an
exception is raised in the ELFHashTable constructor while trying to
parse the struct.
This is fixed by adding the section size to the constructor parameters
of ELFHashTable and constructing an empty table if the length is 0.
Note that not all code paths know the size of the table beforehand, so
it's valid to pass size=None to denote an empty section.
Note that there is still room for errors if the section is e.g. 1 byte
long, but such malformed files should probably just raise an exception.